week3/day1/requirements.txt
attrs==25.3.0
blinker==1.9.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.0
fake2db==0.5.4
Faker==0.7.11
flasgger==0.9.7.1
Flask==3.1.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
MarkupSafe==3.0.2
mistune==3.1.3
packaging==25.0
psycopg2==2.9.10
psycopg2-binary==2.9.10
python-dateutil==2.9.0.post0
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.3
rpds-py==0.25.0
six==1.17.0
tabulate==0.9.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
Werkzeug==3.1.3
